+++ inc/utility.hpp
@@ -1,0 +1,53 @@
+/* SPDX-FileCopyrightText: Copyright (c) 2024 Oğuz Toraman <oguz.toraman@tutanota.com> */
+/* SPDX-License-Identifier: LGPL-3.0-only */
+
+#ifndef UTILITY_HPP
+#define UTILITY_HPP
+
+#include <string>
+#include <ranges>
+#include <concepts>
+#include <algorithm>
+#include <functional>
+
+namespace utility {
+
+/**
+ * @brief Convert any container to string using the string_converter.
+ *
+ * @tparam ContainerType         The type of the container.
+ * @tparam StringConverterType   The type of the string converter callable.
+ *
+ * @param[in] container
+ * @param[in] value_separator    The separator between the values of the container.
+ * @param[in] string_converter   The callable that converts values of the container to string.
+ * 
+ * @returns The container as a string.
+ */
+template <typename ContainerType, typename StringConverterType>
+requires std::ranges::range<ContainerType> && requires (ContainerType c){c.empty(); typename ContainerType::value_type;}
+[[nodiscard]]
+inline std::string to_string(
+    const ContainerType& container,
+    const std::string& value_separator, StringConverterType string_converter)
+{
+    static_assert(
+        std::same_as<std::invoke_result_t<StringConverterType, typename ContainerType::value_type>, std::string>,
+        "StringConverterType must return std::string"
+    );
+    if (container.empty()){
+        return {};
+    }
+    return std::ranges::fold_left(
+        std::ranges::next(std::ranges::begin(container)),
+        std::ranges::end(container),
+        std::invoke(string_converter, *std::ranges::begin(container)),
+        [&](const auto& left, const auto& right){
+            return left + value_separator + std::invoke(string_converter, right);
+        }
+    );
+}
+
+} /* namespace utility */
+
+#endif /* UTILITY_HPP */